+++ engine/geometry/IRaytraceableSurface.cs
@@ -29,6 +29,6 @@ namespace FreedomOfFormFoundation.AnatomyEngine.Geometry
 		/// <param name="ray">
 		///		The Ray object that defines the starting point and direction of the ray to be cast.
 		/// </param>
-		float RayIntersect(Ray ray);
+		double RayIntersect(Ray ray);
 	}
 }
